refactor: Move Input chip with Suggestion Chip
