+++ app/src/main/java/com/infomaniak/swisstransfer/ui/screen/newtransfer/importfiles/components/EmailAddressTextField.kt
@@ -39,59 +39,3 @@ fun EmailAddressTextField(modifier: Modifier) {
 
 }
 
-@Composable
-fun SwissTransferInputChip(
-    modifier: Modifier = Modifier,
-    text: String,
-    onDismiss: () -> Unit,
-) {
-    var enabled by remember { mutableStateOf(false) }
-
-    InputChip(
-        modifier = modifier,
-        border = null,
-        colors = InputChipDefaults.inputChipColors(
-            containerColor = SwissTransferTheme.colors.emailAddressChipColor,
-            labelColor = SwissTransferTheme.colors.onEmailAddressChipColor,
-            selectedLabelColor = SwissTransferTheme.colors.emailAddressChipColor,
-            selectedContainerColor = SwissTransferTheme.colors.onEmailAddressChipColor,
-            selectedTrailingIconColor = SwissTransferTheme.colors.emailAddressChipColor,
-            trailingIconColor = SwissTransferTheme.colors.onEmailAddressChipColor,
-        ),
-        onClick = {
-            onDismiss()
-            enabled = !enabled
-        },
-        label = { Text(text) },
-        selected = enabled,
-        trailingIcon = {
-            IconButton(modifier = Modifier.size(Dimens.IconSize), onClick = {}) {
-                Icon(
-                    modifier = Modifier.size(Dimens.MiniIconSize),
-                    imageVector = AppImages.AppIcons.CrossThick,
-                    contentDescription = "Localized description", // TODO
-                )
-            }
-        },
-    )
-}
-
-@PreviewLightDark
-@Composable
-private fun Preview() {
-    SwissTransferTheme {
-        Surface {
-            Row(
-                Modifier
-                    .padding(Margin.Small)
-                    .horizontalScroll(rememberScrollState()),
-                horizontalArrangement = Arrangement.spacedBy(Margin.Mini)
-            ) {
-                SwissTransferInputChip(text = LoremIpsum(2).values.joinToString(separator = " ")) { }
-                SwissTransferInputChip(text = LoremIpsum(1).values.joinToString(separator = " ")) { }
-                SwissTransferInputChip(text = LoremIpsum(4).values.joinToString(separator = " ")) { }
-                SwissTransferInputChip(text = LoremIpsum(1).values.joinToString(separator = " ")) { }
-            }
-        }
-    }
-}
